requirements.txt
beautifulsoup4==4.11.1
configparser_crypt==1.1.0
google_api_python_client==2.95.0
google_auth_oauthlib==1.0.0
gTTS==2.3.2
playsound==1.3.0
protobuf==3.20.3
PyAudio==0.2.14
pydub==0.25.1
Requests==2.31.0
selenium==4.11.2
undetected_chromedriver==3.5.0
